Added SavedSearchController Modal 'Close' Buttons to renamed to 'Cancel'
